backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.7.2
pdfplumber==0.11.4
python-docx==1.1.2
xhtml2pdf==0.2.16
jinja2==3.1.4
waitress==3.0.1
